Score changes
THe high score table is now more accurate (giving town and dungeon
instead of just "in the fortress" / "in the town").

The score is now calculated in a more complex way.
"Completion" values (0..1) are calculated from the % of level 50
experience, of maximum quest level and of the quests completed
successfully (split into win-track, guardian and others).
They are then weighted individually and summed, scaled so that a
"standard scoring" character with 100% completion gets 1e7 points.
Finally score is scaled by race, extension and personality as
before.