Cleaned up wish parser and unit test.
The wish parser takes a string resembling an item description and builds
an item to fit with fuzzy matching to names - it should make a good guess
at something that is nearly right. Currently it can handle base items,
any legal combination of base item and ego, egos without a base item, and
artifacts. It does not yet handle multiple egos, or understand about
which items should not be wishable.

This version passes all tests, which check for every base item, artifact,
ego, and base item/ego combination with the exception of those where
ambiguity is allowable (an item with the same name as another item, or
an ego item).