Reduce the effectiveness of high movespeed values
Like speed, high movespeed is powerful and the cyber legs can be too
powerful. So the divisor previously used was replaced with a table,
allowing the effect to roll off - the effect of +2 or more to move
speed is reduced. +3 gives the same effect as +2 previously did,
while to get the same effect as +3 previously did you need to max it
at +12 (probably impossible).
A table for negative movement speed penalties has also been added,
though they are not yet used.