Pref file parsing fixes
Failed files now plog() as well as print a message, as the message can't be
seen at startup.
A bug in message.prf uncovered by this was fixed.
The user.prf referred to old races and classes. To avoid future maintenance
the per-race and -class cases were replaced with one line for each of $RACE,
$CLASS, and the new $EXTENSION and $PERSONALITY. This requires parsing the
variables when handling included file names, as well as the existing use in
conditionals. (EQU syntax remains available if you want to add more complex
cases - a specific combination of race and class for example.)