Fix disappearing monsters summoned on death
When a monster is killed, death_special() runs before delete_monster_idx()
does. death_special() may (e.g. for the white horror) summon monsters, and
since the dying monster's space has already been cleared before calling
death_special() they may appear on the same grid as the original monster
(which is often desirable behaviour, when the monster is supposed to be
'transforming into' something else).

If this happens though, delete_monster_idx() would clear the grid again.
This had the effect of a monster which was present in the monster list
but not in the level (you could walk across it, but it would be restored
if you saved and restored the game).

Fixed, by adding a condition to delete_monster_idx() - it won't clear the
grid unless it still contains the original monster.