Fix rogue-like keyset look mode
The key handling in target_set_interactive() was accepting both 'k'
and Ctrl-D to ignore the selection. This meant that 'k' wasn't
accepted as a movement command when using rogue keys. So now 'k'
will only be accepted when using the original keys, and Ctrl-D will
only be accepted when using rogue keys.