Let you know when there is a quest available.
It's easily missable that you can press Q in stores to look for a quest.
While it's on the same help page as all the other new store actions, the
difference is that quests are unexpected - they vary from one shop to
another and over time.

So adding a notification helps make it more discoverable. This is done
with a message at the center bottom whenever you are in a shop that
either has a quest available, or can be used to complete an open quest.

Additionally, when you enter a shop with a new quest available that
wasn't available the last time you visited (or if it was your first
visit), then the usual welcome message gets replaced with one informing
you that the quest is available.