Random names have gender
The first-name list is split into male, female and unisex. Shop owners
have gender assigned when their name is generated (currently those with
a fixed name are always female, though this doesn't affect anything yet).

This is then used to correctly set flags when a shk-monster is built for
fighting.