Shared homes
If a birth option is true (which it is by default), the homes in
all towns have the same shared inventory. (The inventory management
had been called out as annoying, and with such a high capacity (480)
you aren't really losing anything.)

This does mean that town quests for a larger house are no longer
meaningful, but the new per-town home-entry quests are a reasonable
replacement.