Banish-proofing
There are some quests where Banishment aka Extermination can either
make things too easy, or (because the quest needs you to kill a
monster and this is no longer possible) impossible to complete.
So in these cases it is less effective - the last monster on the
level of a home quest can't be banished, and nor can a town monster
that is the last one or not in LOS. Monsters which can't be banished
are teleported away instead.