Fix use-after-free in stats
Stats (clearing) was regenerating the randart list. This caused the
kind list to be reallocated, invalidating pointers into it resulting
in use of unallocated memory (and also making k->max wrong, so saved
files were broken).

This change uses sval 1 if the artifact list is being regenerated,
which is not accurate so this isn't a complete fix.