Limiting of wish power
Wishes from normal play (rather than debug options) are limited.

There are no quest artifacts,  special-generation or zero allocation
probability objects.
There is at most your slot capacity of anything.
There is nothing that's too heavy to carry, unless it's a single item
(not a stack).
There are no stacks that are too expensive, though it won't reject a
single item for price.

There is at most one:
Artifact, but that's not checked for here because artifact creation
does it.
Ego item which is not ammo.
Equippable which is not ammo.